refactor: separated functions and routes
